Extract NullReturnValueBuilder to its own file
